Updated SQL examples in Rf and Lasso code to clean data
SQL examples in rf and lasso dev and deploy did not have predictedCol
being binary or numeric for classification and regression, respectively.
The csv examples had ways to deal with missing data on import.  The csv
import also recognized the regression's independent variable as numeric
whereas the SQL import did not.